+++ cmd/vega/genesis/sign.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"os"
 
-	"code.vegaprotocol.io/go-wallet/wallets"
+	"code.vegaprotocol.io/vegawallet/wallets"
 	"code.vegaprotocol.io/vega/config"
 	"code.vegaprotocol.io/vega/genesis"
 	"code.vegaprotocol.io/vega/logging"