+++ nodewallets/vega/loader.go
@@ -6,9 +6,9 @@ import (
 	"path/filepath"
 	"time"
 
-	"code.vegaprotocol.io/go-wallet/wallet"
-	storev1 "code.vegaprotocol.io/go-wallet/wallet/store/v1"
-	"code.vegaprotocol.io/go-wallet/wallets"
+	"code.vegaprotocol.io/vegawallet/wallet"
+	storev1 "code.vegaprotocol.io/vegawallet/wallet/store/v1"
+	"code.vegaprotocol.io/vegawallet/wallets"
 	"code.vegaprotocol.io/shared/paths"
 	"code.vegaprotocol.io/vega/crypto"
 )