+++ processor/tx_v2.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"fmt"
 
-	wcrypto "code.vegaprotocol.io/go-wallet/crypto"
+	wcrypto "code.vegaprotocol.io/vegawallet/crypto"
 	"code.vegaprotocol.io/protos/commands"
 	commandspb "code.vegaprotocol.io/protos/vega/commands/v1"
 	"code.vegaprotocol.io/vega/libs/crypto"